Title: 실수와 문자 받아 출력하기, Time: 102ms, Memory: 24MB, Status: Passed - CodeTree

+++ 240509/실수와 문자 받아 출력하기/print-number--and-letter.py
@@ -1,1 +1,1 @@
-a=input();b=int(input());c=int(input());print(a);print(f"{b:.2f}");print(f"{c:.2f}");
+a=input();b=float(input());c=float(input());print(a);print(f"{b:.2f}");print(f"{c:.2f}");
